Organizing code a bit better
